Refactor and clean up
- Enfore PositiveInteger on IDiceRoller
- Fix unit tests after moving to value object
- Fix a few warnings
- Add readonly modifiers where needed

+++ Dice/Dice.cs
@@ -39,7 +39,7 @@ namespace DMTools.Die
         /// <returns>Random number between 1 and Sides</returns>
         public int Roll()
         {
-            return _diceRoller.RollDice((int)Sides);
+            return _diceRoller.RollDice(Sides);
         }
 
         public IEnumerable<int> GetResults()
